Fixed Chapter01/03_Taskflow

+++ Chapter01/03_Taskflow/src/main.cpp
@@ -11,7 +11,8 @@ int main()
 
   std::vector<int> items{ 1, 2, 3, 4, 5, 6, 7, 8 };
 
-  auto task = taskflow.for_each_index(0ull, items.size(), 1ull, [&](int i) { printf("%i", items[i]); }).name("for_each_index");
+  auto task =
+      taskflow.for_each_index(0u, static_cast<uint32_t>(items.size()), 1u, [&](int i) { printf("%i", items[i]); }).name("for_each_index");
 
   taskflow.emplace([]() { printf("\nS - Start\n"); }).name("S").precede(task);
   taskflow.emplace([]() { printf("\nT - End\n"); }).name("T").succeed(task);
